+++ examples/data-sources/nxos_ipv4_interface_address/data-source.tf
@@ -1,5 +1,0 @@
-data "nxos_ipv4_interface_address" "example" {
-  vrf          = "default"
-  interface_id = "eth1/10"
-  address      = "24.63.46.49/30"
-}