+++ examples/resources/nxos_ipv4_interface/import-by-identity.tf
@@ -1,7 +1,0 @@
-import {
-  to = nxos_ipv4_interface.example
-  identity = {
-    "vrf" : "<vrf>"
-    "interface_id" : "<interface_id>"
-  }
-}