+++ examples/resources/nxos_ipv4_interface/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_ipv4_interface.example
-  id = "<vrf>,<interface_id>"
-}