+++ examples/resources/nxos_ipv4_interface/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv4_interface.example "<vrf>,<interface_id>"