+++ examples/resources/nxos_ipv4_interface/resource.tf
@@ -1,8 +1,0 @@
-resource "nxos_ipv4_interface" "example" {
-  vrf          = "default"
-  interface_id = "eth1/10"
-  drop_glean   = "disabled"
-  forward      = "disabled"
-  unnumbered   = "unspecified"
-  urpf         = "disabled"
-}