+++ examples/resources/nxos_ipv4_static_route/import-by-identity.tf
@@ -1,7 +1,0 @@
-import {
-  to = nxos_ipv4_static_route.example
-  identity = {
-    "vrf_name" : "<vrf_name>"
-    "prefix" : "<prefix>"
-  }
-}